move: Moved files like model.py, train_model.py, predict_model.py into team_ops/models and updated the env to work with this chage.
