+++ dockerfiles/predict_model.dockerfile
@@ -16,4 +16,4 @@ COPY models/ models/
 
 RUN pip install . --no-cache-dir
 
-ENTRYPOINT ["python", "-u", "team_ops/predict_model.py"]
+ENTRYPOINT ["python", "-u", "team_ops/models/predict_model.py"]
