+++ pyproject.toml
@@ -16,10 +16,11 @@ requires-python = ">=3.11"
 dynamic = ["dependencies", "optional-dependencies"]
 
 [tool.setuptools]
-packages = ["team_ops"]
+packages = ["team_ops", "team_ops.models"]
 
 [tool.setuptools.dynamic]
 dependencies = {file = ["requirements/requirements.txt", "requirements/requirements_manual.txt"]}
+
 
 [tool.setuptools.package-data]
 "team_ops" = ["conf/*.yaml", "conf/experiments/*.yaml"]
